Script for scraping topic titles from url
Read a link from the input. Get all text paragraphs from all the a tags and single out the topics (not any word or phrase!) starting with the letter S. Print the final list. Beware of empty lines.

For example, for this input and letter L:

https://www.who.int/health-topics
the output should be:

["Landslides", "Lassa fever", "Leishmaniasis", "Leprosy", "Lymphatic filariasis"]
The link in the test will be the same one.